ladder: Fix reviewer process
Commit ee7eea33e604d removed reference to one of the teams, but kept the
process as requiring membership in two teams. There is no such
requirement to be a reviewer in multiple teams, the other team was
intended to be the general "reviewer" group. Since we don't have a
dedicated file for all reviewers currently, we don't need to mention the
other team. Make the steps make sense.

Signed-off-by: Joe Stringer <joe@isovalent.com>