add smagnani96 to sig-encryption, wireguard, and ipsec
From the time I started contributing to the Cilium codebase, I've been
focussing on both datapath and control plane for IPSec and Wireguard.
As I plan to continue contributing to this area of expertise, I'd kindly
ask to add myself to these teams, so that I can also provide my
feedback and insights in reviews.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>