Remove inactive reviewers
During the transition of backend tooling, we identified these members
who were in practice not part of the teams although the community repo
stated they are expected to provide reviews for the corresponding teams.
As far as I can tell they no longer primarily focus on Cilium in their
day-to-day development so I don't want to re-impose expectations for
reviews upon them unless they actively signal that they would like to
return to reviewer status for these teams. Thus, remove them for now.

Signed-off-by: Joe Stringer <joe@cilium.io>