members: Add krmeinders to members
Katie Meinders will soon be an active member of the Community team. She
contributes to community content, outreach, and open source engagement around
Cilium and related projects. Adding her to the members ensures she has the
appropriate access and visibility to collaborate on repositories like
cilium.io, and to participate in content and community-related reviews.

Signed-off-by: Donia Chaiehloudj <donia.cld@isovalent.com>