tools: Allow codeowners to be overridden
The linter for generating codeowners so far was mainly focused on
ensuring that each team will request review from other members of the
team, so that teams can self-organize. The merging of the team
membership changes are done through approval and merge from the
committers group on cilium/community repository.

However, the previous iteration of this scripting did not allow
codeowners paths to be overridden in specific cases, such as subprojects
which have different approval patterns.

Update the linter to ensure that each team has an owner, but do not
override existing owners in the tree.

Signed-off-by: Joe Stringer <joe@cilium.io>